chore: create modules folder and move utils.py there
